requirements.txt
# Базовые библиотеки
numpy==1.24.6
opencv-python==4.8.1.78
Pillow==10.0.0
matplotlib==3.8.0
requests==2.31.0

# PyTorch и сопутствующие пакеты
torch==2.1.0
torchvision==0.16.1
torchaudio==2.1.0
torchmetrics==0.11.5

# Альбументации для аугментаций
albumentations==1.3.1

# Для сегментации
segmentation-models-pytorch==0.3.2

# Streamlit для веб-приложения
streamlit==1.28.0

# Для YOLO (Ultralytics)
ultralytics==8.1.0

# Для обработки и визуализации данных
scikit-learn==1.3.0
tqdm==4.66.1
